refactor: remove the anyhow/eyre-style WrapErr/Context feature
WrapErr/Context (attach a message/context to an error) is unused by oxc, the
crate's primary consumer, so drop it to shrink the fork's surface:

- The `WrapErr` trait, the `Context` alias, and the `Result` impl.
- `Report::wrap_err`/`context` and `from_msg`, plus `ContextError` and its four
  dedicated unsafe vtable functions (`context_downcast`, `context_drop_rest`,
  `context_chain_downcast`, `context_chain_drop_rest`) from the type-erasure
  core in report_impl.rs.
- The WrapErr tests (test_context/test_fmt/test_location/test_chain and the
  wrap_err cases in test_source/test_boxed) and the lib.rs WrapErr doc example.

`wrap_err.rs` becomes `wrapper.rs` — it still holds `MessageError`, `BoxedError`,
and `WithSourceCode`, which `Report`'s constructors need. The `fancy` panic hook
(previously `.with_context()`) now folds the panic location into the message and
renders the `Panic` directly; graphical cause-chains were already removed, so the
wrapper chain wasn't being displayed anyway.

The unsafe vtable paths (downcast/drop/source) were checked under `cargo miri`
after the excision. The public API oxc uses is unchanged.
