+++ src/report_impl.rs
@@ -21,7 +21,7 @@ use crate::{
     Diagnostic, Report, ReportHandler, SourceCode,
     chain::Chain,
     ptr::{Mut, Own, Ref},
-    wrap_err::WithSourceCode,
+    wrapper::WithSourceCode,
 };
 
 impl Report {
@@ -123,7 +123,7 @@ impl Report {
     where
         M: Display + Debug + Send + Sync + 'static,
     {
-        use crate::wrap_err::MessageError;
+        use crate::wrapper::MessageError;
         let error: MessageError<M> = MessageError(message);
         let vtable = &ErrorVTable {
             object_drop: object_drop::<MessageError<M>>,
@@ -143,32 +143,8 @@ impl Report {
     }
 
     #[track_caller]
-    pub(crate) fn from_msg<D, E>(msg: D, error: E) -> Self
-    where
-        D: Display + Send + Sync + 'static,
-        E: Diagnostic + Send + Sync + 'static,
-    {
-        let error: ContextError<D, E> = ContextError { msg, error };
-
-        let vtable = &ErrorVTable {
-            object_drop: object_drop::<ContextError<D, E>>,
-            object_ref: object_ref::<ContextError<D, E>>,
-            object_ref_stderr: object_ref_stderr::<ContextError<D, E>>,
-            object_boxed: object_boxed::<ContextError<D, E>>,
-            object_boxed_stderr: object_boxed_stderr::<ContextError<D, E>>,
-            object_downcast: context_downcast::<D, E>,
-            object_drop_rest: context_drop_rest::<D, E>,
-        };
-
-        // Safety: passing vtable that operates on the right type.
-        let handler = Some(crate::report::capture_handler(&error));
-
-        unsafe { Report::construct(error, vtable, handler) }
-    }
-
-    #[track_caller]
     pub(crate) fn from_boxed(error: Box<dyn Diagnostic + Send + Sync>) -> Self {
-        use crate::wrap_err::BoxedError;
+        use crate::wrapper::BoxedError;
         let error = BoxedError(error);
         let handler = Some(crate::report::capture_handler(&error));
 
@@ -209,44 +185,6 @@ impl Report {
         // caller rather than a builtin fat pointer vtable.
         let inner = Own::new(inner).cast::<ErasedErrorImpl>();
         Report { inner }
-    }
-
-    /// Create a new error from an error message to wrap the existing error.
-    ///
-    /// For attaching a higher level error message to a `Result` as it is
-    /// propagated, the [`WrapErr`](crate::WrapErr) extension trait may be more
-    /// convenient than this function.
-    ///
-    /// The primary reason to use `error.wrap_err(...)` instead of
-    /// `result.wrap_err(...)` via the `WrapErr` trait would be if the
-    /// message needs to depend on some data held by the underlying error:
-    pub fn wrap_err<D>(self, msg: D) -> Self
-    where
-        D: Display + Send + Sync + 'static,
-    {
-        let handler = unsafe { self.inner.by_mut().deref_mut().handler.take() };
-        let error: ContextError<D, Report> = ContextError { msg, error: self };
-
-        let vtable = &ErrorVTable {
-            object_drop: object_drop::<ContextError<D, Report>>,
-            object_ref: object_ref::<ContextError<D, Report>>,
-            object_ref_stderr: object_ref_stderr::<ContextError<D, Report>>,
-            object_boxed: object_boxed::<ContextError<D, Report>>,
-            object_boxed_stderr: object_boxed_stderr::<ContextError<D, Report>>,
-            object_downcast: context_chain_downcast::<D>,
-            object_drop_rest: context_chain_drop_rest::<D>,
-        };
-
-        // Safety: passing vtable that operates on the right type.
-        unsafe { Report::construct(error, vtable, handler) }
-    }
-
-    /// Compatibility re-export of `wrap_err` for interop with `anyhow`
-    pub fn context<D>(self, msg: D) -> Self
-    where
-        D: Display + Send + Sync + 'static,
-    {
-        self.wrap_err(msg)
     }
 
     /// An iterator of the chain of source errors contained by this Report.
@@ -566,88 +504,6 @@ where
     }
 }
 
-// Safety: requires layout of *e to match ErrorImpl<ContextError<D, E>>.
-unsafe fn context_downcast<D, E>(e: Ref<'_, ErasedErrorImpl>, target: TypeId) -> Option<Ref<'_, ()>>
-where
-    D: 'static,
-    E: 'static,
-{
-    unsafe {
-        if TypeId::of::<D>() == target {
-            let unerased = e.cast::<ErrorImpl<ContextError<D, E>>>().deref();
-            Some(Ref::new(&unerased._object.msg).cast::<()>())
-        } else if TypeId::of::<E>() == target {
-            let unerased = e.cast::<ErrorImpl<ContextError<D, E>>>().deref();
-            Some(Ref::new(&unerased._object.error).cast::<()>())
-        } else {
-            None
-        }
-    }
-}
-
-// Safety: requires layout of *e to match ErrorImpl<ContextError<D, E>>.
-unsafe fn context_drop_rest<D, E>(e: Own<ErasedErrorImpl>, target: TypeId)
-where
-    D: 'static,
-    E: 'static,
-{
-    unsafe {
-        // Called after downcasting by value to either the D or the E and doing a
-        // ptr::read to take ownership of that value.
-        if TypeId::of::<D>() == target {
-            let unerased = e.cast::<ErrorImpl<ContextError<ManuallyDrop<D>, E>>>().boxed();
-            drop(unerased);
-        } else {
-            let unerased = e.cast::<ErrorImpl<ContextError<D, ManuallyDrop<E>>>>().boxed();
-            drop(unerased);
-        }
-    }
-}
-
-// Safety: requires layout of *e to match ErrorImpl<ContextError<D, Report>>.
-unsafe fn context_chain_downcast<D>(
-    e: Ref<'_, ErasedErrorImpl>,
-    target: TypeId,
-) -> Option<Ref<'_, ()>>
-where
-    D: 'static,
-{
-    unsafe {
-        let unerased = e.cast::<ErrorImpl<ContextError<D, Report>>>().deref();
-        if TypeId::of::<D>() == target {
-            Some(Ref::new(&unerased._object.msg).cast::<()>())
-        } else {
-            // Recurse down the context chain per the inner error's vtable.
-            let source = &unerased._object.error;
-            (vtable(source.inner.ptr).object_downcast)(source.inner.by_ref(), target)
-        }
-    }
-}
-
-// Safety: requires layout of *e to match ErrorImpl<ContextError<D, Report>>.
-unsafe fn context_chain_drop_rest<D>(e: Own<ErasedErrorImpl>, target: TypeId)
-where
-    D: 'static,
-{
-    unsafe {
-        // Called after downcasting by value to either the D or one of the causes
-        // and doing a ptr::read to take ownership of that value.
-        if TypeId::of::<D>() == target {
-            let unerased = e.cast::<ErrorImpl<ContextError<ManuallyDrop<D>, Report>>>().boxed();
-            // Drop the entire rest of the data structure rooted in the next Report.
-            drop(unerased);
-        } else {
-            let unerased = e.cast::<ErrorImpl<ContextError<D, ManuallyDrop<Report>>>>().boxed();
-            // Read out a ManuallyDrop<Box<ErasedErrorImpl>> from the next error.
-            let inner = unerased._object.error.inner;
-            drop(unerased);
-            let vtable = vtable(inner.ptr);
-            // Recursively drop the next error using the same target typeid.
-            (vtable.object_drop_rest)(inner, target);
-        }
-    }
-}
-
 // repr C to ensure that E remains in the final position.
 #[repr(C)]
 pub(crate) struct ErrorImpl<E> {
@@ -656,14 +512,6 @@ pub(crate) struct ErrorImpl<E> {
     // NOTE: Don't use directly. Use only through vtable. Erased type may have
     // different alignment.
     _object: E,
-}
-
-// repr C to ensure that ContextError<D, E> has the same layout as
-// ContextError<ManuallyDrop<D>, E> and ContextError<D, ManuallyDrop<E>>.
-#[repr(C)]
-pub(crate) struct ContextError<D, E> {
-    pub(crate) msg: D,
-    pub(crate) error: E,
 }
 
 type ErasedErrorImpl = ErrorImpl<()>;
